fix : 빌드 점검 및 GMS x

See merge request s14-fintech-finance-sub1/S14P21D102!39

+++ backend/user-service/src/main/java/com/whatsyouretf/userservice/domain/etf/entity/EtfOtherComposition.java
@@ -50,7 +50,7 @@ public class EtfOtherComposition {
 
     /** 시가 */
     @Column(name = "market_value")
-    private Integer marketValue;
+    private Long marketValue;
 
     @Column(name = "created_at")
     @Builder.Default
